Rename status.js as kite-status.js
