+++ lib/ready.js
@@ -10,7 +10,7 @@ const {findClosestGitRepo} = kiteUtils;
 const utils = require('./utils.js');
 const metrics = require('./metrics.js');
 const Login = require('./elements/login.js');
-const StatusItem = require('./elements/status.js');
+const StatusItem = require('./elements/kite-status.js');
 
 // minimum time between showing the same notification, in seconds
 const NOTIFY_DELAY = 60 * 60;
